.gitignore update
Scrapping 'predictive analysis' version of this project until further notice.
This is due to the Dataset that has dates allowing for this kind of analysis, being a filtered subsect without any public documentation on what exactly those filters were. If I am given access to a larger version I may return to this project.